Move to Optional and document
