+++ arrow-libs/optics/arrow-optics/api/arrow-optics.klib.api
@@ -289,7 +289,6 @@ final fun <#A: kotlin/Any?, #B: kotlin/Any?> (arrow.optics/POptional<#A, #A, arr
 final fun <#A: kotlin/Any?, #B: kotlin/Any?> (arrow.optics/POptional<#A, #A, kotlin.collections/List<#B>, kotlin.collections/List<#B>>).arrow.optics.dsl/index(kotlin/Int): arrow.optics/POptional<#A, #A, #B, #B> // arrow.optics.dsl/index|index@arrow.optics.POptional<0:0,0:0,kotlin.collections.List<0:1>,kotlin.collections.List<0:1>>(kotlin.Int){0§<kotlin.Any?>;1§<kotlin.Any?>}[0]
 final fun <#A: kotlin/Any?, #B: kotlin/Any?> (arrow.optics/POptional<#A, #A, kotlin.collections/Set<#B>, kotlin.collections/Set<#B>>).arrow.optics.dsl/at(#B): arrow.optics/POptional<#A, #A, kotlin/Boolean, kotlin/Boolean> // arrow.optics.dsl/at|at@arrow.optics.POptional<0:0,0:0,kotlin.collections.Set<0:1>,kotlin.collections.Set<0:1>>(0:1){0§<kotlin.Any?>;1§<kotlin.Any?>}[0]
 final fun <#A: kotlin/Any?, #B: kotlin/Any?> (arrow.optics/POptional<#A, #A, kotlin.sequences/Sequence<#B>, kotlin.sequences/Sequence<#B>>).arrow.optics.dsl/index(kotlin/Int): arrow.optics/POptional<#A, #A, #B, #B> // arrow.optics.dsl/index|index@arrow.optics.POptional<0:0,0:0,kotlin.sequences.Sequence<0:1>,kotlin.sequences.Sequence<0:1>>(kotlin.Int){0§<kotlin.Any?>;1§<kotlin.Any?>}[0]
-final fun <#A: kotlin/Any?, #B: kotlin/Any?> (arrow.optics/PPrism<#A, #A, #B, #B>).arrow.optics.dsl/filter(kotlin/Function1<#B, kotlin/Boolean>): arrow.optics/PPrism<#A, #A, #B, #B> // arrow.optics.dsl/filter|filter@arrow.optics.PPrism<0:0,0:0,0:1,0:1>(kotlin.Function1<0:1,kotlin.Boolean>){0§<kotlin.Any?>;1§<kotlin.Any?>}[0]
 final fun <#A: kotlin/Any?, #B: kotlin/Any?> (arrow.optics/PTraversal<#A, #A, #B, #B>).arrow.optics.dsl/filter(kotlin/Function1<#B, kotlin/Boolean>): arrow.optics/PTraversal<#A, #A, #B, #B> // arrow.optics.dsl/filter|filter@arrow.optics.PTraversal<0:0,0:0,0:1,0:1>(kotlin.Function1<0:1,kotlin.Boolean>){0§<kotlin.Any?>;1§<kotlin.Any?>}[0]
 final fun <#A: kotlin/Any?, #B: kotlin/Any?> (arrow.optics/PTraversal<#A, #A, arrow.core/NonEmptyList<#B>, arrow.core/NonEmptyList<#B>>).arrow.optics.dsl/filterIndex(kotlin/Function1<kotlin/Int, kotlin/Boolean>): arrow.optics/PTraversal<#A, #A, #B, #B> // arrow.optics.dsl/filterIndex|filterIndex@arrow.optics.PTraversal<0:0,0:0,arrow.core.NonEmptyList<0:1>,arrow.core.NonEmptyList<0:1>>(kotlin.Function1<kotlin.Int,kotlin.Boolean>){0§<kotlin.Any?>;1§<kotlin.Any?>}[0]
 final fun <#A: kotlin/Any?, #B: kotlin/Any?> (arrow.optics/PTraversal<#A, #A, arrow.core/NonEmptyList<#B>, arrow.core/NonEmptyList<#B>>).arrow.optics.dsl/index(kotlin/Int): arrow.optics/PTraversal<#A, #A, #B, #B> // arrow.optics.dsl/index|index@arrow.optics.PTraversal<0:0,0:0,arrow.core.NonEmptyList<0:1>,arrow.core.NonEmptyList<0:1>>(kotlin.Int){0§<kotlin.Any?>;1§<kotlin.Any?>}[0]
@@ -304,8 +303,8 @@ final fun <#A: kotlin/Any?, #B: kotlin/Any?> arrow.optics/Prism(kotlin/Function1
 final fun <#A: kotlin/Any?> (#A).arrow.optics.match/matchUnit(kotlin/Function1<arrow.optics.match/MatchScope<#A, kotlin/Unit>, kotlin/Unit>) // arrow.optics.match/matchUnit|matchUnit@0:0(kotlin.Function1<arrow.optics.match.MatchScope<0:0,kotlin.Unit>,kotlin.Unit>){0§<kotlin.Any?>}[0]
 final fun <#A: kotlin/Any?> (#A).arrow.optics/cons(kotlin.collections/List<#A>): kotlin.collections/List<#A> // arrow.optics/cons|cons@0:0(kotlin.collections.List<0:0>){0§<kotlin.Any?>}[0]
 final fun <#A: kotlin/Any?> (#A).arrow.optics/copy(kotlin/Function1<arrow.optics/Copy<#A>, kotlin/Unit>): #A // arrow.optics/copy|copy@0:0(kotlin.Function1<arrow.optics.Copy<0:0>,kotlin.Unit>){0§<kotlin.Any?>}[0]
+final fun <#A: kotlin/Any?> (arrow.optics/POptional.Companion).arrow.optics/filter(kotlin/Function1<#A, kotlin/Boolean>): arrow.optics/POptional<#A, #A, #A, #A> // arrow.optics/filter|filter@arrow.optics.POptional.Companion(kotlin.Function1<0:0,kotlin.Boolean>){0§<kotlin.Any?>}[0]
 final fun <#A: kotlin/Any?> (arrow.optics/POptional<#A, #A, kotlin/String, kotlin/String>).arrow.optics.dsl/index(kotlin/Int): arrow.optics/POptional<#A, #A, kotlin/Char, kotlin/Char> // arrow.optics.dsl/index|index@arrow.optics.POptional<0:0,0:0,kotlin.String,kotlin.String>(kotlin.Int){0§<kotlin.Any?>}[0]
-final fun <#A: kotlin/Any?> (arrow.optics/PPrism.Companion).arrow.optics/filter(kotlin/Function1<#A, kotlin/Boolean>): arrow.optics/PPrism<#A, #A, #A, #A> // arrow.optics/filter|filter@arrow.optics.PPrism.Companion(kotlin.Function1<0:0,kotlin.Boolean>){0§<kotlin.Any?>}[0]
 final fun <#A: kotlin/Any?> (arrow.optics/PTraversal<#A, #A, kotlin/String, kotlin/String>).arrow.optics.dsl/filterIndex(kotlin/Function1<kotlin/Int, kotlin/Boolean>): arrow.optics/PTraversal<#A, #A, kotlin/Char, kotlin/Char> // arrow.optics.dsl/filterIndex|filterIndex@arrow.optics.PTraversal<0:0,0:0,kotlin.String,kotlin.String>(kotlin.Function1<kotlin.Int,kotlin.Boolean>){0§<kotlin.Any?>}[0]
 final fun <#A: kotlin/Any?> (arrow.optics/PTraversal<#A, #A, kotlin/String, kotlin/String>).arrow.optics.dsl/index(kotlin/Int): arrow.optics/PTraversal<#A, #A, kotlin/Char, kotlin/Char> // arrow.optics.dsl/index|index@arrow.optics.PTraversal<0:0,0:0,kotlin.String,kotlin.String>(kotlin.Int){0§<kotlin.Any?>}[0]
 final fun <#A: kotlin/Any?> (kotlin.collections/List<#A>).arrow.optics/snoc(#A): kotlin.collections/List<#A> // arrow.optics/snoc|snoc@kotlin.collections.List<0:0>(0:0){0§<kotlin.Any?>}[0]
